Time: 12 ms (85.77%), Space: 8.6 MB (38.39%) - LeetHub

+++ 451-sort-characters-by-frequency/451-sort-characters-by-frequency.cpp
@@ -1,17 +1,17 @@
 class Solution {
 public:
-    static bool cmp(pair<char,long long> p1, pair<char,long long> p2){
+    static bool cmp(pair<char,int> p1, pair<char,int> p2){
         return p1.second>p2.second;
     }
     string frequencySort(string s) {
-        map<char,long long>hm;
+        map<char,int>hm;
         for(char c: s) hm[c]++;
-        vector<pair<char,long long>> vec;
+        vector<pair<char,int>> vec;
         for(auto x: hm) vec.push_back({x.first,x.second});
         sort(vec.begin(),vec.end(),cmp);
         string ans;
         for(auto x: vec)
-            for(long long i=0;i<x.second;i++) 
+            for(int i=0;i<x.second;i++) 
                 ans+=x.first;
         
         return ans;
